+++ src/chapters/algorithmic-thinking/what-is-algorithm.js
@@ -3,11 +3,10 @@ export const whatIsAlgorithmChapter = {
   title: 'What is an Algorithm?',
   sectionId: 'algorithmic-thinking',
   previousChapterId: null,
-  content: `## What is an Algorithm?
-
+  content: `
 An algorithm is a step-by-step procedure or a set of rules designed to perform a specific task or solve a particular problem. Algorithms are the foundation of computer science and programming, providing structured approaches to problem-solving.
 
-TBD
+<iframe width="1120" height="630" src="https://www.youtube.com/embed/fkIvmfqX-t0?si=wj8A7L1JKZ9GqyGk" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe>
 
 ## Key Characteristics of Algorithms
 
@@ -17,7 +16,8 @@ TBD
 - **Output**: An algorithm produces one or more outputs.
 - **Effectiveness**: Each step must be basic enough to be carried out by a person using only pencil and paper.
 
-TBD
+Ever tried making a peanut butter sandwich with vague instructions like “put the peanut butter on the bread”? In this video, we show what happens when algorithms aren’t definite—turns out, your sandwich ends up a mess! Algorithms need to be clear and exact, or you might find yourself spreading peanut butter on the outside of the bag. 
+<iframe width="1120" height="630" src="https://www.youtube.com/embed/cDA3_5982h8?si=V0HOZsxmQCqfhU77" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe>
 
 ## Why Algorithms Matter
 
@@ -28,16 +28,12 @@ Algorithms are essential in computer science and software development for severa
 - They enable us to analyze and compare different approaches to solving problems
 - They help us understand the limitations and capabilities of computing
 
-TBD
-
 ## Examples of Common Algorithms
 
 - **Sorting algorithms**: Bubble sort, merge sort, quicksort
 - **Search algorithms**: Linear search, binary search
 - **Graph algorithms**: Breadth-first search, depth-first search, Dijkstra's algorithm
 - **Dynamic programming algorithms**: Fibonacci sequence, knapsack problem
-- **Machine learning algorithms**: Linear regression, k-means clustering, neural networks
-
-TBD`,
+`,
   exercise: null
 };
